tools/pdf-to-md/pyproject.toml
[project]
name = "pdf-to-md"
version = "0.1.0"
description = "PDF to Markdown conversion tool using marker-pdf"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0.0",
    "marker-pdf~=0.2.0",
    "transformers==4.45.2",
    "pyyaml>=6.0.0",
]

[project.scripts]
pdf-to-md = "pdf_to_md:cli"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
